requirements.txt
appdirs==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
bleach==3.3.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
entrypoints==0.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lockfile==0.12.2
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==0.6.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
packaging==20.3
pandocfilters==1.4.3
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.17
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.22.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.14.0
terminado==0.9.3
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
